Lagt til kommentarer i Room_for_personRepository Lagt til Rolle i Room_for_personRepository Endret variabler til camelCase i Room Lagt til rolle i Room_for_person.java Laget Room_for_personController.java

+++ Spring_JPA_N-g/src/main/java/com/nag/spring_jpa_ng/repository/Room_for_personRepository.java
@@ -4,11 +4,24 @@ import com.nag.spring_jpa_ng.model.Room_for_person;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface Room_for_personRepository extends JpaRepository<Room_for_person, Long> {
-    List <Room_for_person> findByRoomId(Long roomId);
+    // Hente alle room_for_person basert på roomID
+    List <Room_for_person> findByRoomId(int roomId);
+
+    //Hente alle Room_for_person basert på persinID
     List <Room_for_person> findByPersonId(Long personId);
-    List <Room_for_person> findByRoomIdAndPersonId(Long roomId, Long personId);
-    List <Room_for_person> findByRoomIdAndPersonIdAndRole(Long roomId, Long personId, String role);
-    List <Room_for_person> findByRoomIdAndRole(Long roomId, String role);
+
+    // Hente alle Room_for_person basert på både roomID og personID
+    List <Room_for_person> findByRoomIdAndPersonId(int roomId, Long personId);
+
+    // Hente alle Room_for_person basert på roomID, personID og rolle
+    List <Room_for_person> findByRoomIdAndPersonIdAndRole(int roomId, Long personId, String role);
+
+    // Hente alle Room_for_person basert på roomID og rolle
+    List <Room_for_person> findByRoomIdAndRole(int roomId, String role);
+
+    //Hente alle Room_for_person basert på personID og rolle
     List <Room_for_person> findByPersonIdAndRole(Long personId, String role);
+
+    // Hente alle Room_for_person med spesifisert rolle
     List <Room_for_person> findByRole(String role);
 }
